HeroBonus: split BonusList
