+++ lib/spells/TargetCondition.cpp
@@ -16,6 +16,7 @@
 #include "../battle/CBattleInfoCallback.h"
 #include "../battle/Unit.h"
 #include "../bonuses/BonusParams.h"
+#include "../bonuses/BonusList.h"
 
 #include "../serializer/JsonSerializeFormat.h"
 #include "../VCMI_Lib.h"
