+++ scripting/lua/api/BonusSystem.cpp
@@ -11,6 +11,7 @@
 
 #include "BonusSystem.h"
 
+#include "../../../lib/bonuses/BonusList.h"
 #include "../../../lib/bonuses/HeroBonus.h"
 #include "../../../lib/bonuses/IBonusBearer.h"
