+++ test/mock/mock_BonusBearer.h
@@ -10,6 +10,7 @@
 
 #pragma once
 
+#include "../../lib/bonuses/BonusList.h"
 #include "../../lib/bonuses/HeroBonus.h"
 #include "../../lib/bonuses/IBonusBearer.h"
